+++ examples/examples-maximilian/11-Mixing/11-Mixing.ino
@@ -21,7 +21,7 @@ void setup() {//some inits
     maximilian.begin(cfg);
 }
 
-void play(double *output) {
+void play(float *output) {
     myOutputs.stereo(myOsc.noise(),myStereoOutput,(myAutoPanner.sinewave(1)+1)/2);//Stereo, Quad or 8 Channel. Specify the input to be mixed, the output[numberofchannels], and the pan (0-1,equal power).
     output[0]=myStereoOutput[0];//When working with mixing, you need to specify the outputs explicitly
     output[1]=myStereoOutput[1];//
